hw-10: Fixed bugs. Added sweetAlert for function delete book.

+++ hw10-library-react/library-react/src/main/js/components/Book.js
@@ -1,11 +1,24 @@
 import React from "react";
 import FormUpdateBook from "./FormUpdateBook";
+import swal from 'sweetalert';
 
 export default class Book extends React.Component {
 
     handleDeleteBook = () => {
-        window.confirm("Are you sure you wish to delete the book?") &&
-        this.props.handleDeleteBook(this.props.book);
+        swal({
+            title: "Are you sure?",
+            text: "Are you sure that you want to delete this book?",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+        })
+            .then((willDelete) => {
+                if (willDelete) {
+                    this.props.handleDeleteBook(this.props.book);
+                } else {
+                    swal.close();
+                }
+            });
     };
 
     render() {
@@ -19,6 +32,8 @@ export default class Book extends React.Component {
                 <td>
                     <div className="input-group">
                         <FormUpdateBook attributes={this.props.attributes}
+                                        authors={this.props.authors}
+                                        genres={this.props.genres}
                                         book={this.props.book}
                                         handleUpdateBook={this.props.handleUpdateBook}/>
                         <button onClick={this.handleDeleteBook} className="btn btn-outline-danger btn-sm" type="button">
